Moved compiler-specific code out of the completion contributor and into a HaxeCompilerServices class.
